switch closed/open hash names
