+++ src/AlgoList.js
@@ -16,8 +16,8 @@ export const algoMap = {
 	BST: ['Binary Search Tree', algos.BST],
 	Heap: ['Heaps / PQs', algos.Heap],
 	SkipList: ['SkipList', algos.SkipList],
-	ClosedHash: ['HashMap (Probing)', algos.ClosedHash],
-	OpenHash: ['HashMap (Chaining)', algos.OpenHash],
+	OpenHash: ['HashMap (Probing)', algos.OpenHash],
+	ClosedHash: ['HashMap (Chaining)', algos.ClosedHash],
 	SplayTree: ['SplayTree', algos.SplayTree],
 	AVL: ['AVL', algos.AVL],
 	BTree: ['2-4 Tree', algos.BTree],
