+++ src/algo/index.js
@@ -7,7 +7,7 @@ export { default as BoyerMoore } from './BoyerMoore';
 export { default as BruteForce } from './BruteForce';
 export { default as BubbleSort } from './BubbleSort';
 export { default as CircularlyLinkedList } from './CircularlyLinkedList';
-export { default as ClosedHash } from './ClosedHash';
+export { default as OpenHash } from './OpenHash';
 export { default as CocktailSort } from './CocktailSort';
 export { default as DFS } from './DFS';
 export { default as LCS } from './LCS';
@@ -23,7 +23,7 @@ export { default as Kruskals } from './Kruskals';
 export { default as LSDRadix } from './LSDRadix';
 export { default as LinkedList } from './LinkedList';
 export { default as MergeSort } from './MergeSort';
-export { default as OpenHash } from './OpenHash';
+export { default as ClosedHash } from './ClosedHash';
 export { default as Prims } from './Prims';
 export { default as QueueArray } from './QueueArray';
 export { default as QueueLL } from './QueueLL';
